breaking-change(profile): image update profil

+++ src/api/v1/controllers/user.controller.js
@@ -51,8 +51,11 @@ class UserController {
 
   static async update(req, res, next) {
     try {
-      const image = await cloudinary.uploader.upload(req.file.path);
-      fs.unlinkSync(req.file.path);
+      let image;
+      if (req.file) {
+        image = await cloudinary.uploader.upload(req.file.path);
+        fs.unlinkSync(req.file.path);
+      }
 
       if (req.user) {
         await UserService.updateUserBiodata(
